POKEMON-193 Add-Spark-microservice-to-transfer-data-from-HBase-to-Doris-by-Controller

+++ backend/hadoop/spark-service/src/main/java/com/pokemoney/hadoop/spark/model/TransactionAnalysisModel.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * The simple analysis model.
  */
@@ -12,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TransactionAnalysisModel {
     private Long id;
-    private Double sumMoney;
+    private BigDecimal sumMoney;
     private Integer categoryId;
     private Integer typeId;
     private Integer timeAnalysisId;
